+++ digits_3d/inference_test.py
@@ -8,7 +8,7 @@ from digit_classify import  digit_classify
 
 if __name__ == '__main__':
     folder_path = 'training_data'
-    file_name = 'stroke_7_0006.csv'
+    file_name = 'stroke_9_0052.csv'
     to_load = os.path.join(folder_path, file_name)
     digit = pd.read_csv(to_load, header=None)  # load digit
     digit = digit.to_numpy()                   # convert to numpy array
